requirements.txt
# 核心Web框架
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# 认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# 数据库
sqlalchemy>=2.0.23
##sqlite3
chromadb>=0.4.19

# 大模型&OpenClaw
openclaw>=0.5.0
dashscope>=1.14.0  # 通义千问SDK
python-dotenv>=1.0.0

# 图像处理
opencv-python>=4.8.1.78
pillow>=10.1.0
pytesseract>=0.3.10
numpy>=1.26.2

# 其他工具
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
loguru>=0.7.2